fix: type of setGroupSize function

+++ include/gwmodelpp/GWRBasic.h
@@ -687,7 +687,7 @@ public:     // Implement IParallelizable
 public:     // Implement IGwmParallelOpenmpEnabled
     void setOmpThreadNum(const int threadNum) override { mOmpThreadNum = threadNum; }
     void setGPUId(const int gpuId) override { mGpuId = gpuId; };
-    void setGroupSize(const double size) override { mGroupLength = size; };
+    void setGroupSize(const size_t size) override { mGroupLength = size; };
 
 protected:
